Refactor methods in RemoteConnections and PasqalCloud for new connections
